✅[CHORE] : 타이머 다이얼 로그 오탈자 수정 및 종속성 최소화 작업
